Move schema utils to `metabase/lib`
