+++ frontend/src/metabase/lib/schema/index.js
@@ -1,0 +1,1 @@
+export * from "./schema";
